Add preview files for iOS and watchOS
To fix SwiftUI preview issues with two widget targets
